utils/ordered_dict: don't use compat code for Python 3.8+
